pyproject.toml
[tool.poetry]
name = "dbre-be"
version = "0.1.0"
description = ""
authors = ["limsu1015 <tngur1015@gmail.com>"]
packages = [
    { include = "dbre_BE" }
]
package-mode = true

[tool.poetry.dependencies]
python = "3.12.6"
django = ">=5.1.5,<6.0.0"
djangorestframework-simplejwt = ">=5.4.0,<6.0.0"
psycopg2-binary = "^2.9.10"
poetry = "1.8.5"
drf-spectacular = "^0.28.0"
django-allauth = "^65.3.1"
cryptography = "^44.0.0"
django-redis = "^5.4.0"
twilio = "^9.4.4"
types-requests = "^2.32.0.20241016"

[build-system]
requires = ["poetry-core>=1.8.5"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
requests = "^2.32.3"
portone-server-sdk = "^0.10.0"
gunicorn = "^23.0.0"
django-cors-headers = "^4.6.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"
django-stubs = "^5.1.2"
python-dateutil = "^2.9.0.post0"

[tool.black]
line-length = 88  # 코드 한 줄의 최대 길이 (기본값: 88)
target-version = ["py312"]  # Python 버전 명시

[tool.isort]
profile = "black"  # Black과 호환되도록 설정
atomic = true
force_grid_wrap = 0
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
multi_line_output = 3
use_parentheses = true

[tool.mypy]
python_version = "3.12"  # Python 버전 명시
plugins = ["mypy_django_plugin.main"]  # Django 플러그인 활성화

# 타입 검사 옵션
disallow_untyped_defs = true  # 모든 함수에 타입 힌트를 요구
ignore_missing_imports = true  # 누락된 import 경고 무시
warn_unused_ignores = true  # 사용되지 않는 ignore 경고 표시
warn_return_any = true  # 반환값이 Any일 경우 경고

# 특정 모듈에 대한 예외 처리 (선택 사항)
[[tool.mypy.overrides]]
module = ["django", "django.db.models"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "dbre_BE.settings.local"
ignore_missing_model_attributes = true  # 모델 속성 누락 무시
strict_settings = true  # settings.py의 엄격한 타입 검사 활성화